refactor(P4b-2): extract Escalation mixin verbatim from Executor
Provider-call lifecycle (single + escalating, sync + stream + responses-
API), error/retry classification, and corresponding audit/metering
emission move into Inference::Executor::Escalation. Verbatim move; no
semantic changes.

Executor.rb: 2402 → 1827 lines. escalation.rb: 593 lines.
ClassLength rubocop:disable removed (no longer triggered).
